Ban user

+++ src/app/api/user/comment/route.ts
@@ -25,22 +25,42 @@ export async function POST(req: NextRequest) {
 			return NextResponse.json({ error: "Not allowed" }, { status: 401 });
 		}
 
+		const userValidate = await prisma.user.findUnique({
+			where: {
+				id: user.id,
+			},
+		});
+
+		if (!userValidate) {
+			return NextResponse.json({ error: "Not allowed" }, { status: 401 });
+		}
+
+		if (userValidate?.role === "banned") {
+			return NextResponse.json({ error: "You are banned!" }, { status: 403 });
+		}
+
 		const comments = await prisma.comment.findMany();
 
 		if (comments.length > 25) {
 			await prisma.comment.delete({
 				where: {
 					id: comments.at(0)?.id,
-					comment: comments.at(0)?.comment,
+					text: comments.at(0)?.text,
 				},
 			});
 		}
 
 		await prisma.comment.create({
 			data: {
-				comment,
-				authorId: user.id,
-				author: user.username,
+				text: comment,
+				author: {
+					connect: {
+						id: user.id,
+					},
+				},
+			},
+			include: {
+				author: true,
 			},
 		});
 
@@ -55,7 +75,17 @@ export async function POST(req: NextRequest) {
 
 export async function GET() {
 	try {
-		const comments = await prisma.comment.findMany();
+		const comments = await prisma.comment.findMany({
+			include: {
+				author: {
+					select: {
+						username: true,
+						role: true,
+						about: true,
+					},
+				},
+			},
+		});
 
 		comments.forEach(async (validComment) => {
 			const validUser = await prisma.user.findUnique({
